--- examples/operator/basic/basic.go
@@ -1,6 +1,0 @@
-package main
-
-type BasicModel struct {
-	Number int    `json:"numField"`
-	String string `json:"stringField"`
-}
