+++ examples/operator/basic/reconciler/main.go
@@ -108,3 +108,8 @@ func main() {
 		panic(fmt.Errorf("error running operator: %w", err))
 	}
 }
+
+type BasicModel struct {
+	Number int    `json:"numField"`
+	String string `json:"stringField"`
+}
