+++ examples/operator/basic/watcher/main.go
@@ -114,3 +114,8 @@ func main() {
 		panic(fmt.Errorf("error running operator: %w", err))
 	}
 }
+
+type BasicModel struct {
+	Number int    `json:"numField"`
+	String string `json:"stringField"`
+}
