+++ examples/operator/opinionated/run.sh
@@ -12,10 +12,10 @@ sh "${ROOT_DIR}/examples/create_cluster.sh"
 # TODO: run in cluster?
 case "$1" in
   watcher)
-    go run watcher.go opinionated.go --kubecfg="${HOME}/.kube/config"
+    go run watcher/main.go --kubecfg="${HOME}/.kube/config"
     ;;
   reconciler)
-    go run reconciler.go opinionated.go --kubecfg="${HOME}/.kube/config"
+    go run reconciler/main.go --kubecfg="${HOME}/.kube/config"
     ;;
   *)
     echo "unknown option: $1" >&2
