+++ examples/operator/opinionated/watcher/main.go
@@ -17,6 +17,15 @@ import (
 	"github.com/grafana/grafana-app-sdk/resource"
 )
 
+// These are for pretty printing in the logs
+const (
+	nc     = "\033[0m"
+	red    = "\033[1;31m"
+	green  = "\033[1;32m"
+	yellow = "\033[1;33m"
+	blue   = "\033[1;34m"
+)
+
 func main() {
 	log.SetPrefix("\033[0;37m")
 	kubeCfgFile := flag.String("kubecfg", "", "kube config path")
@@ -126,4 +135,9 @@ func main() {
 
 	// Run the controller (will block until stopCh receives a message or is closed)
 	op.Run(stopCh)
+}
+
+type OpinionatedModel struct {
+	Number int    `json:"numField"`
+	String string `json:"stringField"`
 }
